% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{bdcr_species_search}
\alias{bdcr_species_search}
\title{Buscar información taxonómica de una especie en BIODATACR}
\usage{
bdcr_species_search(name, rows = 10)
}
\arguments{
\item{name}{Character. Nombre científico (puede ser sinónimo o parcial).}

\item{rows}{Entero. Número máximo de resultados. Default 10.}
}
\value{
`tibble` con columnas: `name`, `guid`, `commonName`,
  `scientificName`, `rank`, `kingdom`, `family`, `occurrenceCount`.
}
\description{
Buscar información taxonómica de una especie en BIODATACR
}
\examples{
\dontrun{
bdcr_species_search("Panthera onca")
}
}
