% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_outline}
\alias{cr_outline}
\title{Costa Rica national boundary}
\format{
An `sf` object with 1 row and the following columns:
\describe{
  \item{GID_0}{Country code (CRI)}
  \item{COUNTRY}{Country name}
  \item{geometry}{MULTIPOLYGON geometry}
}
}
\source{
\url{https://gadm.org/}
}
\usage{
cr_outline
}
\description{
An `sf` object with the national boundary of Costa Rica from GADM.
}
\keyword{datasets}
