% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{bdcr_count}
\alias{bdcr_count}
\title{Contar registros de ocurrencia para un taxón}
\usage{
bdcr_count(taxon)
}
\arguments{
\item{taxon}{Character. Nombre científico del taxón (e.g. `"Panthera onca"`).}
}
\value{
Entero con el número total de registros.
}
\description{
Contar registros de ocurrencia para un taxón
}
\examples{
\dontrun{
bdcr_count("Panthera onca")
}
}
