% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{bdcr_count_batch}
\alias{bdcr_count_batch}
\title{Contar registros para múltiples taxones}
\usage{
bdcr_count_batch(taxa, wait = 1)
}
\arguments{
\item{taxa}{Character vector. Nombres científicos.}

\item{wait}{Numeric. Segundos de pausa entre consultas. Default 1.}
}
\value{
`tibble` con columnas `taxon` y `n_records`.
}
\description{
Contar registros para múltiples taxones
}
