% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{bdcr_occurrences}
\alias{bdcr_occurrences}
\title{Descargar registros de ocurrencia de BIODATACR}
\usage{
bdcr_occurrences(taxon, rows = 100, start = 0)
}
\arguments{
\item{taxon}{Character. Nombre científico del taxón.}

\item{rows}{Entero. Número máximo de registros. Default 100.}

\item{start}{Entero. Registro inicial para paginación. Default 0.}
}
\value{
`tibble` con campos de ocurrencia.
}
\description{
Descargar registros de ocurrencia de BIODATACR
}
